refactor: to use useRelativeTime to display the og content created time.
