--- src/components/ui/custom/relative-time-formatter.tsx
@@ -1,14 +1,0 @@
-import { useFormatter, useNow } from "next-intl"
-import React from "react"
-
-export const RelativeTimeFormatter: React.FC<{ createdAt: string }> = ({ createdAt }) => {
-  const now = useNow({ updateInterval: 1000 * 10 })
-  const format = useFormatter()
-  const formatRelativeTime = () => {
-    const createdAtDate = new Date(createdAt)
-
-    return format.relativeTime(createdAtDate, now)
-  }
-
-  return <>{formatRelativeTime()}</>
-}
